pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "winclaw"
version = "1.0.17"
description = "WinClaw - 轻量级 Windows AI 桌面智能体"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "litellm>=1.40",
    "openai>=1.30",
    "mss>=9.0",
    "Pillow>=10.0",
    "rich>=13.0",
    "tomli>=2.0; python_version < '3.11'",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "apscheduler>=3.10",
    "aiosqlite>=0.20",
]

[project.optional-dependencies]
gui = [
    "PySide6>=6.7",
    "qasync>=0.27",
    "keyring>=25.0",
    "pynput>=1.7",
]
automation = [
    "pyautogui>=0.9",
    "pywinauto>=0.6",
    "pyperclip>=1.8",
    "winotify>=1.1",
]
browser = [
    "playwright>=1.44",
]
voice = [
    "openai-whisper>=20240930",
    "pyttsx3>=2.90",
    "sounddevice>=0.4.6",
    "scipy>=1.11",
    "numpy>=1.24",
]
ocr = [
    "rapidocr-onnxruntime>=1.3",
]
mcp = [
    "mcp>=1.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
]
all = [
    "winclaw[gui,automation,browser,voice,ocr,mcp,dev]",
]

[project.scripts]
winclaw = "src.app:main"
winclaw-gui = "src.ui.gui_app:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "optional: marks tests requiring optional dependencies",
]
